wallet: move GetAccount across all backends
